address round-3 review: %lu cast + platform-gate NULL check
Two cosmetic fixes from /review 290 (v3):

- exceptions.c: cast current_cpu to (unsigned long) for the %lu
  PSCI-off banner. Variadic strict-aliasing — current_cpu is
  uint32_t and %lu expects unsigned long; AAPCS made it work in
  practice but the cast is correct.

- sched.c: gate the sched_diag_idle_loops NULL guard behind
  PLATFORM_HAS_NC_MEMORY. On other platforms the symbol is a
  BSS array (address always non-NULL) and GCC's -Werror=address
  rejects `if (arr)` as "always true". The NULL case only
  applies when ncmem_alloc could have failed (NC-backed
  platforms).

Verified: make test (QEMU ARM64) passes.

+++ kernel/arch/arm64/exceptions.c
@@ -180,9 +180,13 @@ static void handle_page_fault(struct trap_frame *tf, uint64_t esr, uint64_t far,
                   "(PSCI-offing the boot CPU means no supervisor "
                   "can resurrect it).\n");
     } else {
+        /* Cast to unsigned long for %lu — current_cpu is uint32_t
+         * but our uart_printf is variadic; passing a 32-bit value
+         * to %lu reads extra bytes from the arg register on
+         * strict implementations. */
         uart_printf("\nCPU %lu offlining via PSCI — remaining CPUs "
                     "continue (see #216 for auto-resurrection plan).\n",
-                    current_cpu);
+                    (unsigned long)current_cpu);
     }
     psci_cpu_off();
 #endif
